requirements.txt
flask
discord.py
python-dotenv
requests
waitress
beautifulsoup4
pillow
asyncio
aiohttp